requirements.txt
asgiref==3.2.10
Django==3.0.8
django-environ==0.4.5
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
mysql
pillow==8.3.2
djangorestframework==3.12.4
django_filter==2.4.0
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0